+++ mwcp/__init__.py
@@ -11,7 +11,9 @@ from mwcp.parser import Parser
 from mwcp.file_object import FileObject
 from mwcp.registry import (
     register_entry_points, register_parser_directory, register_parser_package,
-    iter_parsers, get_parser_descriptions, set_default_source, clear_default_source)
+    iter_parsers, get_parser_descriptions, set_default_source, clear_default_source,
+    ParserNotFoundError
+)
 from mwcp.reporter import Reporter
 from mwcp.resources import techanarchy_bridge
 from mwcp.dispatcher import Dispatcher, UnableToParse, UnidentifiedFile